pyproject.toml
[project]
name = "impulse"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.9",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
